requirements.txt
requests
watchdog
python-telegram-bot==11.1.0
selenium